add missing cast without it, the player would lock up on occasion
git-svn-id: http://svn.mythtv.org/svn/trunk@370 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/NuppelVideoPlayer.cpp
@@ -891,9 +891,9 @@ void NuppelVideoPlayer::GetFrame(int onlyvideo)
         {
 	  /* we didnt get a known frametype! 
 	     so move down one char and look again... */
-            ringBuffer->Seek(1-FRAMEHEADERSIZE, SEEK_CUR);
+            ringBuffer->Seek((long long)1-FRAMEHEADERSIZE, SEEK_CUR);
 
-            if (ringBuffer->Read(&frameheader, FRAMEHEADERSIZE) 
+            if (ringBuffer->Read(&frameheader, FRAMEHEADERSIZE)
                 != FRAMEHEADERSIZE)
             {
                 eof = 1;
